backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
httpx==0.27.0
pillow>=10.0.0
requests>=2.31.0
